Sync release with dev

+++ src/app/pages/tasks/healthkit/services/health-questionnaire-processor.service.ts
@@ -11,6 +11,7 @@ import { getMilliseconds } from 'src/app/shared/utilities/time'
 import { LogService } from '../../../../core/services/misc/log.service'
 import { QuestionnaireProcessorService } from 'src/app/pages/questions/services/questionnaire-processor/questionnaire-processor.service'
 import { HealthkitService } from './healthkit.service'
+import { NotificationService } from 'src/app/core/services/notifications/notification.service'
 
 interface ProcessingProgress {
   stage: 'validation' | 'processing' | 'compression' | 'complete'
@@ -31,8 +32,9 @@ export class HealthQuestionnaireProcessorService extends QuestionnaireProcessorS
     kafka: KafkaService,
     private logger: LogService,
     private healthkit: HealthkitService,
+    notifications: NotificationService
   ) {
-    super(schedule, kafka)
+    super(schedule, kafka, notifications)
   }
 
   async getUnsentHealthkitCount(): Promise<number> {
@@ -81,7 +83,8 @@ export class HealthQuestionnaireProcessorService extends QuestionnaireProcessorS
               .then(() => this.kafka.sendAllFromCache())
               .then((sendResult) => {
                 const allSent = sendResult && Array.isArray(sendResult.failedKeys) && sendResult.failedKeys.length === 0 && !sendResult.cancelled
-                if (!allSent) {
+                const hasHealthkitFailedKeys = sendResult && Array.isArray(sendResult.failedKeys) && sendResult.failedKeys.filter((key) => key.startsWith(SchemaType.HEALTHKIT)).length > 0
+                if (!allSent && hasHealthkitFailedKeys) {
                   worker.terminate()
                   return Promise.reject(new Error('Not all HealthKit data could be sent.'))
                 }
@@ -137,7 +140,8 @@ export class HealthQuestionnaireProcessorService extends QuestionnaireProcessorS
           .then(() => this.kafka.sendAllFromCache())
           .then((sendResult) => {
             const allSent = sendResult && Array.isArray(sendResult.failedKeys) && sendResult.failedKeys.length === 0 && !sendResult.cancelled
-            if (!allSent) {
+            const hasHealthkitFailedKeys = sendResult && Array.isArray(sendResult.failedKeys) && sendResult.failedKeys.filter((key) => key.startsWith(SchemaType.HEALTHKIT)).length > 0
+            if (!allSent && hasHealthkitFailedKeys) {
               return Promise.reject(new Error('Not all HealthKit data could be sent.'))
             }
             return this.healthkit.setUploadReadyFlag(false)
